+++ golem-worker-service-base/src/service/mod.rs
@@ -12,13 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod api_definition_validator;
-pub mod api_deployment;
 pub mod component;
 pub mod worker;
+pub mod worker_gateway;
 
 pub mod http;
-mod worker_gateway;
 
 pub fn with_metadata<T, I, K, V>(request: T, metadata: I) -> tonic::Request<T>
 where
