+++ golem-worker-service-base/src/service/worker_gateway/mod.rs
@@ -1,2 +1,4 @@
 pub mod api_definition;
 pub mod api_definition_lookup;
+pub mod api_definition_validator;
+pub mod api_deployment;
