+++ golem-worker-service/src/api/api_deployment.rs
@@ -7,8 +7,8 @@ use golem_worker_service_base::api::ApiEndpointError;
 use golem_worker_service_base::api::{ApiDeployment, ApiDeploymentRequest};
 use golem_worker_service_base::api_definition;
 use golem_worker_service_base::api_definition::{ApiDefinitionId, ApiSiteString};
-use golem_worker_service_base::service::api_definition::ApiDefinitionIdWithVersion;
-use golem_worker_service_base::service::api_deployment::ApiDeploymentService;
+use golem_worker_service_base::service::worker_gateway::api_definition::ApiDefinitionIdWithVersion;
+use golem_worker_service_base::service::worker_gateway::api_deployment::ApiDeploymentService;
 use poem_openapi::param::{Path, Query};
 use poem_openapi::payload::Json;
 use poem_openapi::*;
